homeassistant/requirements.txt
yamllint>=1.37.0
requests
python-dotenv
